fix: #59 eventseat 패키지 구조 변경, 쿼리 dsl 추가
