+++ build.gradle
@@ -47,4 +47,8 @@ subprojects {
     test {
         useJUnitPlatform()
     }
+
+    clean {
+        delete file('src/main/generated') // 인텔리제이 Annotation processor 생성물 생성위치
+    }
 }
